support extract parameter marker segment in AbstractSQLStatement.
